Making it clear that the extension mechanism has no magic
The implication with the current dreamcode was that under the hood, somewhere magical, extending hoodie was actually implicitly associating the extensions to a subproperty called "dreamcode".

It should allow you to extend hoodie arbitrarily. So, I can have a plugin that adds both "posts", "votes" as properties immediately under hoodie.